// src/home/types/technology-types.ts
export enum TechnologyTypes {
    P5 = 'p5.js',
    THREEJS = 'Three.js',
    REACT = 'React.js',
    OPEN_FRAMEWORKS = "OpenFrameworks",
    PROJECTORS = "projectors",
    CSS = "S/CSS",
    D3 = "d3.js",
    HTML = "HTML",
    WEB_SCRAPING = "web scraping",
    ARDUINO = "Arduino",
    GEO_JSON = "GeoJSON",
    STEPPER_MOTORS = "stepper motors",
}